pyproject.toml
[project]
name = "pae-auth"
version = "0.1.0"
description = ""
authors = [
    {name = "srestrero",email = "srestreporo@unal.edu.co"}
]
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.29.0,<0.30.0)",
    "pydantic[email] (>=2.9.1,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==3.2.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)",
    "prometheus-client (==0.20.0)",
    "opentelemetry-distro (==0.45b0)",
    "opentelemetry-instrumentation-fastapi (==0.45b0)",
    "opentelemetry-instrumentation-logging (==0.45b0)",
    "opentelemetry-exporter-otlp (==1.24.0)"
]

[tool.poetry]
packages = [{include = "*", from = "src"}]

[tool.poetry.group.dev.dependencies]
commitizen = "^4.8.2"
pre-commit = "^4.2.0"
poethepoet = "^0.34.0"
faker = "^26.0.0"
pytest = "^8.3.2"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.poe.tasks]
dev = "uvicorn main:app --reload --app-dir src --host 0.0.0.0 --port 8000 --reload"
test = "pytest"
db-generate = "alembic revision --autogenerate"
db-migrate = "alembic upgrade head"
